feat(Spinner and toasts): added synfusion toasts and custom spinner in login, signup, new entry pages

+++ Components/Cards/EntryJournalCard.razor.cs
@@ -1,15 +1,22 @@
 ﻿using AuroraJournalingApp.Models;
 using AuroraJournalingApp.Services;
+using Microsoft.AspNetCore.Components;
 using Microsoft.Maui.Controls;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+
+
 namespace AuroraJournalingApp.Components.Cards
 {
+
     public partial class EntryJournalCard(JournalService service)
     {
+        [Inject]
+        public NavigationManager nav { get; set; }
+
         Journal todayEntry=new();
         protected override async Task OnInitializedAsync()
         {
@@ -20,7 +27,32 @@ namespace AuroraJournalingApp.Components.Cards
 
         public  async Task deleteAsync()
         {
-             await service.DeleteJournal(todayEntry.JournalId);
+            if (todayEntry == null) return;
+            try
+            {
+                LoadingService.Show();
+                await service.DeleteJournal(todayEntry.JournalId);
+                await ToastService.ShowSuccess("Journal deleted successfully");
+                todayEntry = null;
+
+
+                StateHasChanged();
+            }
+            catch
+            {
+                await ToastService.ShowError("failed to delete journal");
+            }
+            finally
+            {
+                LoadingService.Hide();
+            }
+
+           
+
+          
+           
+
+
 
         }
